requirements.txt
fastapi
uvicorn
pillow
python-multipart
torch
timm
numpy
git+https://github.com/huggingface/transformers.git
